refactor: remove views, add folders for components
