+++ src/Layout/ComputationMenu/index.ts
@@ -2,4 +2,4 @@ export * from "./computation-menu-actions";
 export * from "./computation-menu-reducer";
 export * from "./computation-menu-state";
 export * from "./ComputationMenu";
-export * from "./ComputationMenuContainer";
+export * from "./ConnectedComputationMenu";
