+++ src/Layout/Presentation/SeatsPerParty/index.ts
@@ -1,0 +1,1 @@
+export * from "./SeatsPerParty";
