+++ src/Layout/Presentation/SingleDistrict/SingleDistrict.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import ReactTable from "react-table";
-import { toSum } from "../Utilities/ReduceUtilities";
-import { DistrictResult, PartyResult } from "../../../computation";
+import { DistrictResult, PartyResult } from "../../../computation/computation-models";
+import { toSum } from "../../../utilities/reduce";
 
 export interface SingleDistrictProps {
     districtResults: DistrictResult[];
