+++ src/Layout/Presentation/SingleDistrict/index.ts
@@ -1,0 +1,1 @@
+export * from "./SingleDistrict";
