--- src/Layout/Presentation/Views/index.ts
@@ -1,6 +1,0 @@
-export * from "./DistrictOverview";
-export * from "./ElectionOverview";
-export * from "./SeatDistribution";
-export * from "./SeatsPerParty";
-export * from "./SingleDistrict";
-export * from "./LevellingSeatOverview";
