+++ src/Layout/Presentation/index.ts
@@ -1,2 +1,9 @@
 export * from "./Presentation";
 export * from "./ConnectedPresentation";
+export * from "./DistrictOverview";
+export * from "./ElectionOverview";
+export * from "./LevellingSeatOverview";
+export * from "./RemainderQuotients";
+export * from "./SeatDistribution";
+export * from "./SeatsPerParty";
+export * from "./SingleDistrict";
