+++ src/utilities/reduce.ts
